deleted entries page

+++ src/main.js
@@ -242,19 +242,59 @@ ipcMain.on('get_locations', async (event, formData) => {
 });
 
 ipcMain.on('deleteInventoryEntry', async (event, id) => {
-  try{
-    console.log("calling deleteInventoryEntry...");
-    console.log(id);
-    const result = await userEntries.deleteInventoryEntry(id);
+  try {
+      const response = await dialog.showMessageBox({
+          type: 'question',
+          buttons: ['Cancel', 'Delete'],
+          defaultId: 0,
+          title: 'Confirm Delete',
+          message: 'Are you sure you want to delete this entry?',
+      });
+      if (response.response === 1) {
+          console.log("calling deleteInventoryEntry...", id);
+          const result = await userEntries.deleteInventoryEntry(id);
+          console.log("result of deleteInventoryEntry: ", result);
+          event.reply('deleteInventoryEntry_Response', { response: 1});
+      } else {
+          console.log("Deletion cancelled by user");
+          event.reply('deleteInventoryEntry_Response', { response: 0 });
+      }
+  } catch (error) {
+      console.error('Error in deleteInventoryEntry:', error);
+      event.reply('deleteInventoryEntry_Response', { response: -1, error: error.message });
+  }
+});
 
-    console.log("result of deleteInventoryEntry: ", result);
 
-    event.reply('deleteInventoryEntry_Response', result);
-  }catch(error){
+/*
+ipcMain.on('deleteInventoryEntry', async (event, id) => {
+  try {
+    // Show a confirmation dialog before proceeding
+    const response = await dialog.showMessageBox({
+      type: 'question',
+      buttons: ['Cancel', 'Delete'],
+      defaultId: 0,
+      title: 'Confirm Delete',
+      message: 'Are you sure you want to delete this entry?',
+    });
+    // Check if the user clicked 'Delete'
+    if (response.response === 1) {
+      console.log("calling deleteInventoryEntry...");
+      console.log(id);
+      const result = await userEntries.deleteInventoryEntry(id);
+      console.log("result of deleteInventoryEntry: ", result);
+      event.reply('deleteInventoryEntry_Response', result);
+    } else {
+      // User cancelled the operation
+      event.reply('deleteInventoryEntry_Response', result);
+    }
+  } catch (error) {
     console.error('Error in deleteInventoryEntry:', error);
     event.reply('deleteInventoryEntry_Response', { error: error.message });
   }
 });
+*/
+
 
 ipcMain.on('update_Inventory_Entry', async (event, formData) => {
   try{
